iio: adc: ad7616: Move software mode configuration to spi file
Software mode configuration was moved in spi file because it is only
available when spi is used.
